commit for fixing recipe with new nav
