+++ .idea/shelf/Uncommitted_changes_before_Update_at_27-05-2025_11_43_[Changes]/shelved.patch
@@ -1,0 +1,12 @@
+Index: .idea/shelf/Uncommitted_changes_before_Checkout_at_18-05-2025_21_24__Changes_.xml
+===================================================================
+diff --git a/.idea/shelf/Uncommitted_changes_before_Checkout_at_18-05-2025_21_24__Changes_.xml b/.idea/shelf/Uncommitted_changes_before_Checkout_at_18-05-2025_21_24__Changes_.xml
+deleted file mode 100644
+--- a/.idea/shelf/Uncommitted_changes_before_Checkout_at_18-05-2025_21_24__Changes_.xml	(revision 2f4b9e677d747549c953e1cf003bf458c0e7777a)
++++ /dev/null	(revision 2f4b9e677d747549c953e1cf003bf458c0e7777a)
+@@ -1,4 +0,0 @@
+-<changelist name="Uncommitted_changes_before_Checkout_at_18-05-2025_21_24_[Changes]" date="1747596260540" recycled="true" deleted="true">
+-  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Uncommitted_changes_before_Checkout_at_18-05-2025_21_24_[Changes]/shelved.patch" />
+-  <option name="DESCRIPTION" value="Uncommitted changes before Checkout at 18-05-2025 21:24 [Changes]" />
+-</changelist>
+\ No newline at end of file
