+++ .idea/shelf/Uncommitted_changes_before_Update_at_27-05-2025_11_43__Changes_.xml
@@ -1,0 +1,4 @@
+<changelist name="Uncommitted_changes_before_Update_at_27-05-2025_11_43_[Changes]" date="1748339041316" recycled="true" deleted="true">
+  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Uncommitted_changes_before_Update_at_27-05-2025_11_43_[Changes]/shelved.patch" />
+  <option name="DESCRIPTION" value="Uncommitted changes before Update at 27-05-2025 11:43 [Changes]" />
+</changelist>
