+++ .idea/shelf/Uncommitted_changes_before_Update_at_27-05-2025_12_02__Changes_.xml
@@ -1,0 +1,4 @@
+<changelist name="Uncommitted_changes_before_Update_at_27-05-2025_12_02_[Changes]" date="1748340137748" recycled="true" deleted="true">
+  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Uncommitted_changes_before_Update_at_27-05-2025_12_02_[Changes]/shelved.patch" />
+  <option name="DESCRIPTION" value="Uncommitted changes before Update at 27-05-2025 12:02 [Changes]" />
+</changelist>
