+++ .idea/shelf/Uncommitted_changes_before_Update_at_27-05-2025_12_04__Changes_.xml
@@ -1,0 +1,4 @@
+<changelist name="Uncommitted_changes_before_Update_at_27-05-2025_12_04_[Changes]" date="1748340267908" recycled="true" deleted="true">
+  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Uncommitted_changes_before_Update_at_27-05-2025_12_04_[Changes]/shelved.patch" />
+  <option name="DESCRIPTION" value="Uncommitted changes before Update at 27-05-2025 12:04 [Changes]" />
+</changelist>
