+++ app/build.gradle.kts
@@ -63,7 +63,7 @@ dependencies {
     //firebase dependencies
     implementation(platform(libs.firebase.bom))
     implementation(libs.firebase.analytics)    //vi bruger ikke analytics til noget
-    implementation ("com.google.firebase:firebase-auth:23.2.1")
+    implementation ("com.google.firebase:firebase-auth:23.2.1") //firebase auth
     //viewmodel
     implementation(libs.androidx.lifecycle.viewmodel.compose)
     //navigation
