+++ app/src/main/java/com/example/gronkokken/ui/pages/klimaplan.kt
@@ -1,31 +1,20 @@
-package com.example.gronkokken.ui.theme
+package com.example.gronkokken.ui.pages
 
-import android.R.attr.contentDescription
-import android.R.attr.text
 import com.example.gronkokken.R
 import android.os.Bundle
-import android.util.Log
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material.icons.filled.ArrowBack
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.ExperimentalMaterial3Api
@@ -45,35 +34,30 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.example.gronkokken.components.Firestore
-import com.example.gronkokken.components.UserViewModel
-import com.example.gronkokken.ui.theme.GronKokkenTheme
+import androidx.navigation.NavHostController
+import com.example.gronkokken.repository.Firestore
 import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.firestore.FirebaseFirestore
-import kotlinx.coroutines.tasks.await
 
 
-class ClimaPlan : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            ClimaPlanScreen()
-        }
-    }
-}
+//class ClimaPlan : ComponentActivity() {
+//    override fun onCreate(savedInstanceState: Bundle?) {
+//        super.onCreate(savedInstanceState)
+//        setContent {
+//            ClimaPlanScreen()
+//        }
+//    }
+//}
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun ClimaPlanScreen(userViewModel: UserViewModel = viewModel()) {
+fun ClimaPlanScreen(userViewModel: com.example.gronkokken.repository.UserViewModel = viewModel(), navHostController: NavHostController) {
     val robotoBold = FontFamily(Font(R.font.roboto))
     var startpunkt by remember { mutableStateOf("") }
     var slutpunkt by remember { mutableStateOf("") }
@@ -94,7 +78,7 @@ fun ClimaPlanScreen(userViewModel: UserViewModel = viewModel()) {
             TopAppBar(
                 title = {},
                 navigationIcon = {
-                    IconButton(onClick = { /* Gå tilbage */ }) {
+                    IconButton(onClick = { navHostController.popBackStack() }) {
                         Icon(
                             painter = painterResource(id = R.drawable.baseline_arrow_back_ios_24),
                             contentDescription = "Tilbage"
